// src/Moryx.Orders.Web/app/src/app/api/models/report-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConfirmationType } from '../models/confirmation-type';
export interface ReportModel {
  comment?: string | null;
  confirmationType?: ConfirmationType;
  failureCount?: number;
  successCount?: number;
  userIdentifier?: string | null;
}